[Matthew D'Alonzo] reorganized files
