+++ up_tck_python/examples/up_client_socket_python/test_service.py
