+++ up_tck_python/test_agent/testagent.py
@@ -1,0 +1,37 @@
+# -------------------------------------------------------------------------
+
+# Copyright (c) 2023 General Motors GTO LLC
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+# SPDX-FileType: SOURCE
+# SPDX-FileCopyrightText: 2023 General Motors GTO LLC
+# SPDX-License-Identifier: Apache-2.0
+
+# -------------------------------------------------------------------------
+
+from abc import ABC, abstractmethod
+
+class TestAgent(ABC):
+    
+    @abstractmethod
+    def receive_from_tm(self, listener) -> None:
+        '''
+        Receives commands from Test Manager and processes it..
+        @param listener: Callback that is called when data is received from Test Manager.
+        '''
+        None
